% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMSTCross.R
\docType{data}
\name{CMSTCross}
\alias{CMSTCross}
\title{CMST Sample Cross object}
\format{
object of `cross`
}
\source{
`SimCrossCausal`
}
\usage{
CMSTCross
}
\description{
Created with `SimCrossCausal()`
}
\keyword{datasets}
