% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCommonQtls.R
\name{GetCommonQtls}
\alias{GetCommonQtls}
\title{Get common QTLs for phenotypes}
\usage{
GetCommonQtls(
  cross,
  pheno1,
  pheno2,
  thr = 3,
  peak.dist = 5,
  addcov1 = NULL,
  addcov2 = NULL,
  intcov1 = NULL,
  intcov2 = NULL
)
}
\arguments{
\item{cross}{object of class `cross`}

\item{pheno1}{first phenotype column number or character string name}

\item{pheno2}{second phenotype column number or character string name; if
more than one, then all phenotypes will be tested against `pheno1`}

\item{thr}{LOD threshold}

\item{peak.dist}{maximal peak distance to be considered the same peak (in
cM)}

\item{addcov1, addcov2}{additive covariates for first and second phenotype,
respectively}

\item{intcov1, intcov2}{interactive covariates for first and second
phenotype, respectively}
}
\description{
Perform joint QTL mapping for phenotypes with marginal LOD peak positions
higher than LOD threshold and within set distance of each other
}
\examples{

# data(CMSTCross) should work here but fails on R.check
set.seed(987654321)
CMSTCross <- SimCrossCausal(n.ind = 100, 
  len = rep(100, 3), n.mar = 101,
  beta = rep(0.5, 2), add.eff = 1, dom.eff = 0, 
  sig2.1 = 0.4, sig2.2 = 0.1, eq.spacing = FALSE, 
  cross.type = "bc", normalize = TRUE)
CMSTCross <- calc.genoprob(CMSTCross, step = 1)
commqtls <- GetCommonQtls(CMSTCross, 
                          pheno1 = "y1", 
                          pheno2 = "y3",
                          thr = 3,
                          peak.dist = 5,
                          addcov1 = NULL, 
                          addcov2 = NULL, 
                          intcov1 = NULL, 
                          intcov2 = NULL)
commqtls

}
\references{
Chaibub Neto E, Broman AT, Keller MP, Attie AD, Zhang B, Zhu J,
Yandell BS, Causal model selection hypothesis tests in systems genetics.
Genetics (in review).
}
\keyword{utilities}
