% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotperm.R
\docType{data}
\name{hotperm1}
\alias{hotperm1}
\title{Example object for hotperm results}
\format{
A list with the following components:
\describe{
  \item{max.N}{Matrix of maximum hotspot sizes for different LOD thresholds.}
  \item{max.N.window}{Matrix of smoothed maximum hotspot sizes (if window smoothing is applied).}
  \item{max.lod.quant}{Matrix of maximum LOD scores for different quantiles.}
}
}
\usage{
hotperm1
}
\description{
The `hotperm1` object is an example output from the `hotperm` function. It contains the results of permutation tests for hotspot size thresholds.
This object is used to demonstrate the functionality of the `hotperm` package.
}
\examples{
\dontrun{
data(hotperm1)
summary(hotperm1)
}
}
\keyword{datasets}
