% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include.hotspots.R
\name{include.hotspots}
\alias{include.hotspots}
\title{Generate hotspots for the simulated examples in the manuscript.}
\usage{
include.hotspots(
  cross,
  hchr,
  hpos,
  hsize,
  Q.eff,
  latent.eff,
  lod.range.1,
  lod.range.2,
  lod.range.3,
  res.var = 1,
  n.pheno,
  init.seed
)
}
\arguments{
\item{cross}{cross object}

\item{hchr}{character string for chromosome}

\item{hpos}{position on chromosome}

\item{hsize}{size}

\item{Q.eff}{Q effect}

\item{latent.eff}{latent effect}

\item{lod.range.1}{lod range 1}

\item{lod.range.2}{lod range 2}

\item{lod.range.3}{lod range 3}

\item{res.var}{residual variance}

\item{n.pheno}{number of phenotypes}

\item{init.seed}{initial seed}
}
\description{
Generate hotspots for the simulated examples in the manuscript.
}
