% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neqtl.R
\name{neqtl}
\alias{neqtl}
\title{neqtl.R Ported from http://github.com/kbroman/neqtl}
\usage{
neqtl(sigpos.out, chr, pos, win = 5)
}
\arguments{
\item{sigpos.out}{}

\item{chr}{}

\item{pos}{}

\item{win}{}
}
\description{
neqtl.R Ported from http://github.com/kbroman/neqtl
}
