% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{parallel.qtlhot}
\alias{parallel.qtlhot}
\title{Code for parallelizing R/qtlhot.}
\usage{
parallel.qtlhot(x, data = 1, ..., dirpath = ".")
}
\arguments{
\item{x}{phase of parallel processing (1,2,3)}

\item{data}{index for parallel processing (1,2,...)}

\item{\dots}{additional arguments passed along}

\item{dirpath}{directory path as character string}
}
\description{
Code for parallelizing R/qtlhot. See installed parallel directory for proper
use. There is apparently an S3 parallel method, so doc has to be as shown
below, even though it is called as parallel.qtlhot.
}
\seealso{
\code{\link[qtl]{read.cross}}
}
\author{
Brian S Yandell and Elias Chaibub Neto
}
\keyword{utilities}
