% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simhot.R
\name{sim.null.cross}
\alias{sim.null.cross}
\title{Generates a "null dataset" cross}
\usage{
sim.null.cross(
  chr.len = rep(400, 16),
  n.mar = 185,
  n.ind = 112,
  type = "bc",
  n.pheno = 6000,
  latent.eff = 1.5,
  res.var = 1,
  init.seed = 92387475
)
}
\arguments{
\item{chr.len}{}

\item{n.mar}{}

\item{n.ind}{}

\item{type}{}

\item{n.pheno}{}

\item{latent.eff}{}

\item{res.var}{}

\item{init.seed}{}
}
\description{
Generates a "null dataset" cross
}
