% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcmst.R
\docType{data}
\name{CMSTCross}
\alias{CMSTCross}
\title{Simulate a Cross Object for Causal Model Selection}
\format{
A `cross` object with the following components:
\describe{
  \item{geno}{Genotype probabilities for each marker.}
  \item{pheno}{Phenotype data for each individual.}
  \item{map}{Genetic map of the markers.}
}
}
\usage{
CMSTCross
}
\description{
The `CMSTCross` object is a simulated cross used for causal model selection hypothesis tests.
It is generated using the `SimCrossCausal` function and processed with `calc.genoprob`.
}
\examples{
\dontrun{
set.seed(987654321)
CMSTCross <- SimCrossCausal(n.ind = 100, 
                            len = rep(100, 3), n.mar = 101,
                            beta = rep(0.5, 2), add.eff = 1, dom.eff = 0, 
                            sig2.1 = 0.4, sig2.2 = 0.1, eq.spacing = FALSE, 
                            cross.type = "bc", normalize = TRUE)
CMSTCross <- calc.genoprob(CMSTCross, step = 1)
save(CMSTCross, file = "CMSTCross.RData", compress = TRUE)
}
}
\keyword{datasets}
