% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMSTCross.R
\name{CMSTCross}
\alias{CMSTCross}
\title{MST Sample Cross object}
\source{
`SimCrossCausal`
}
\description{
Created with `SimCrossCausal()`.
}
\details{
`CMSTCross` is created with `SimCrossCausal()`
and used in several examples.
}
\examples{
data(CMSTCross)
}
\keyword{datasets}
