% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotperm1.R
\docType{data}
\name{hotperm1}
\alias{hotperm1}
\title{Hotspot Permutation Sample}
\format{
object of class `hotperm`
}
\source{
`hotperm`
}
\usage{
hotperm1
}
\description{
Created with `hotperm()`
}
\keyword{datasets}
