% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotperm1.R
\name{hotperm1}
\alias{hotperm1}
\title{Hotspot Permutation Sample}
\source{
`SimCrossCausal`
}
\description{
Created with `hotperm()`.
}
\details{
`hotperm1` is created with `hotperm()`
and used in several examples.
}
\examples{
data(hotperm1)
}
\keyword{datasets}
