% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simhot.R
\name{include.hotspots}
\alias{include.hotspots}
\title{Generate hotspots for the simulated examples in the manuscript.}
\usage{
include.hotspots(
  cross,
  hchr,
  hpos,
  hsize,
  Q.eff,
  latent.eff,
  lod.range.1,
  lod.range.2,
  lod.range.3,
  res.var = 1,
  n.pheno,
  init.seed
)
}
\arguments{
\item{cross}{}

\item{hchr}{}

\item{hpos}{}

\item{hsize}{}

\item{Q.eff}{}

\item{latent.eff}{}

\item{lod.range.1}{}

\item{lod.range.2}{}

\item{lod.range.3}{}

\item{res.var}{}

\item{n.pheno}{}

\item{init.seed}{}
}
\description{
Generate hotspots for the simulated examples in the manuscript.
}
