% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neqtl.R
\name{neqtl}
\alias{neqtl}
\title{neqtl.R Ported from http://github.com/kbroman/neqtl}
\usage{
neqtl(sigpos.out, chr, pos, win = 5)
}
\arguments{
\item{sigpos.out}{position(s) of max}

\item{chr}{chromosome as character}

\item{pos}{position on chromosome}

\item{win}{window width in cM}
}
\description{
neqtl.R Ported from http://github.com/kbroman/neqtl
}
