% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.null.cross.R
\name{sim.null.cross}
\alias{sim.null.cross}
\title{Generates a "null dataset" cross}
\usage{
sim.null.cross(
  chr.len = rep(400, 16),
  n.mar = 185,
  n.ind = 112,
  type = "bc",
  n.pheno = 6000,
  latent.eff = 1.5,
  res.var = 1,
  init.seed = 92387475
)
}
\arguments{
\item{chr.len}{vector with length of chromosomes}

\item{n.mar}{number of markers}

\item{n.ind}{number of individuals}

\item{type}{cross type as character}

\item{n.pheno}{number of phenotypes}

\item{latent.eff}{latent effect}

\item{res.var}{residual variance}

\item{init.seed}{initial seed}
}
\description{
Generates a "null dataset" cross
}
